remove core (#118)
We have no plans to make our packages usable outside of this repo in the short or medium term. So put everything in internal/node which allows a bunch of simplification. I'm also deleting any interfaces where we only have one implementation and no mocks of it, since those don't need to be interfaces in that case and a lot of times it's like a sub-type of something in the same package so we can start making things internal. I believe a lot of these were interfaces to get around the import dependency loops that were happening. Also removed some types that were unnecessary like JSONState since all the JSON patch stuff doesn't actually need to be stateful, it can just be helper functions that the relevant structs (like fileDB) calls into. On that note, I think fileDB is a bad name, it's following from hdb / habitat db naming, but it's not really a database, just a JSON blob that is persisted, so I want this to be like a cohesive state package that persists that "state" of the habitat node. Also doing some architectural thinking about what a "node" is (how many users, repos etc.) on the side, so we should make this package super usable because it seems prone to change. The basic thing we need is a patch-able JSON schema and persistence of that, so I'm trying to carve it down to just that.

Follow-ups:
This thing is upside down because we should have packages like reverse_proxy, process etc. and then node is actually something that imports from all of those, because it is the service that puts a bunch of things together. So the next step is to move all of those type definitions next to where the implementation is (i.e. type ReverseProxyRule struct { ... } needs to be in reverse_proxy next to the implementation rather than in internal/node/state) and then node state is just importing those.